cmake: Add support for having CMAKE_BUILD_TYPE be the empty string
Add support for having CMAKE_BUILD_TYPE be the empty string. Setting
CMAKE_BUILD_TYPE to the empty string is valid use, so we should not
override it when it is set as such.

Also, when a multi-config generator is in use, and CMAKE_BUILD_TYPE is
set to the empty string, we will accidentally be dynamically changing
the value of CMAKE_BUILD_TYPE. This breaks the build because as the
docs say, this value is a static value and should not be dynamically
changed:

"This statically specifies what build type

https://cmake.org/cmake/help/v3.17/variable/CMAKE_BUILD_TYPE.html

This is of the same nature as
https://github.com/OpenAMP/libmetal/pull/108

Signed-off-by: Sebastian Bøe <sebastian.boe@nordicsemi.no>