app: rpc_demod: replace copy_from_shbuf with metal_io_block_read
After a toolchain upgrade, the OpenAMP rpc_demod application would
fail with a bus error on AArch64 Linux. This was because the call to
copy_from_shbuf() was now inlined by the optimizer and the compiler
generated code that used 8-byte copies ("ldp x10, x11 [x4, #16]" followed
by the corresponding store instruction). The bus error happened when
the address in x4 (the index) was 4-byte aligned, but not 8-byte aligned.

This would not be a problem with normal user-space memory, but the memory
used for the shared buffer between the remote processor and Linux is
mapped as device memory. On AArch64, device memory is strictly ordered
and the alignment and width of accesses must match.

Replace copy_from_shbuf() with a call to metal_io_block_read(), which
is intended to do this sort of thing, and also makes this a better
example of how to use OpenAMP.

Return error only on a negative return value from copy_from_shbuf in
rpmsg_endpoint_cb.

Signed-off-by: Ben Levinsky <ben.levinsky@xilinx.com>