version: add generated define for version
For memory footprint reason it can be useful for applications
to use defines during the build pre-processing step.

Add a generated version_def.h file that will be accessible in the
the install directory ./build/bin/usr/local/include/openamp/.

Application will be able to use the version.h or the version_def.h
depending on its need.

Signed-off-by: Arnaud Pouliquen <arnaud.pouliquen@foss.st.com>