rpmsg: let rpmsg_virtio_get_tx_buffer() always return idx in host side
In rpmsg host side, the tx buffer index is not assigned when this buffer
is obtained from the reclaimer list, this commit let
rpmsg_virtio_get_tx_buffer() always get correct idx.

Note: the idx in rpmsg host side is not used, so this commit is just
an improvement and not fix any issue.

Signed-off-by: Guiding Li <liguiding1@xiaomi.com>
Signed-off-by: Bowen Wang <wangbowen6@xiaomi.com>