Docs: enable readthedocs rendering (for PRs)
The Readthedocs pull-request integration is nice to have for PR review.
So run the whole OpenAMP-docs build but use the PR's commit

This also requires a ReadTheDocs project to install its integration for
this repo. The project should enable "Build pull request for this project".
It is currently in the Advanced Settings panel.

The project will also run for the main branch but that should be
ignored and only the main project URL should be shown outside of PRs.

To this end the sub-module project names will use "-prs" in the name.

Signed-off-by: Bill Mills <bill.mills@linaro.org>