remoteproc: Assign ops and priv to rproc struct in init by default
These need to be set in the rproc specific init() callback currently,
instead set them before calling the callback to make doing that in the
callback optional.

They can still be overwritten if needed in the init() callback but
if this is not needed as in some simple cases, then the callback is
not needed. Like other functions here, make this optional if no
additional modifications to the rproc struct are needed.

Signed-off-by: Andrew Davis <afd@ti.com>