lib: rpmsg: replace strncpy with internal safe_strcpy
The strncpy function does not ensure that the destination string is
null-terminated. To address this issue, replace strncpy with the
internal safe_strcpy() function, which guarantees null-termination of the
destination string but also access only in buffer memory ranges.

Note: (void)safe_strcpy(...) indicates that the return value is
intentionally ignored.

Signed-off-by: Arnaud Pouliquen <arnaud.pouliquen@foss.st.com>