CI: Do Zephyr build tests on known good and latest versions
Allow Zephyr testing to either use locked known good values
or the very latest versions.

This brings open-amp to parity with libmetal.  Open-amp did not have the
SDK URL discovery code so that was added here as part of this.

The known good versions are best for PR checking as if the build fails
it is almost always the PR itself that broke it.

The latest version is good for periodically checking compatibility with
the very latest Zephyr changes.

For now we run both on pushes and PRs.

We also run main against the latest zephyr check weekly as a look ahead.

Signed-off-by: Bill Mills <bill.mills@linaro.org>