lib/rpmsg: move the ept->release_cb out of the lock region
Add lock inside the rpmsg_ept_inc/decref and move the ept->release_cb
out of the lock region because the destruct callback do not need
hold the lock.
And if the lock is spinlocked based implementation, the new
implementation also can avoid the dead lock issue if there is context
switch in the ept->release_cb.

Signed-off-by: Bowen Wang <wangbowen6@xiaomi.com>