lib/remoteproc: seperate the vdev and vring bitmap
Linux do not alloc the notifyid for the virtio device and use
one idr alloc notifyid for the vring.
OpenAMP use one bitmap alloc the notifyid for both the virtio device
and the vring.

Both the Linux and OpenAMP will assgin back the alloced notifyid
to the resource table and check this notifyid is used before.
This may cause the notifyid is different to the linux alloced
and cause remoteproc init error in OpenAMP.

For example, Linux alloc the vring notifyid: 0, 1 and assgin back to
the resource table, OpenAMP must alloc the same notifyid for vring,
but the notifyid 0 already be used by the virtio device, so OpenAMP
will init failed at the handle_vdev_rsc() with error code:
-RPROC_ERR_RSC_TAB_NP.

Signed-off-by: Bowen Wang <wangbowen6@xiaomi.com>